pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cpa-tax-advisory"
version = "1.0.0"
description = "CPA Tax Advisory Platform"
requires-python = ">=3.10"
dependencies = [
    "langchain==0.1.0",
    "langchain-openai==0.0.2",
    "openai==1.0.0",
    "python-dotenv==1.0.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "pypdf==3.17.0",
    "pdfplumber==0.10.3",
    "pytesseract==0.3.10",
    "Pillow==10.1.0",
    "pdf2image==1.16.3",
    "numpy==1.26.2",
    "matplotlib==3.7.0",
    "scikit-learn==1.3.0",
    "joblib==1.3.0",
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "jinja2==3.1.3",
    "python-dateutil==2.8.2",
    "typing-extensions==4.8.0",
    "pyotp==2.9.0",
    "bcrypt==4.1.2",
    "python-multipart==0.0.6",
    "reportlab==3.6.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
test = [
    "pytest==7.4.4",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.25.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
